requirements.txt
blinker==1.8.2
click==8.1.7
contourpy==1.2.1
cycler==0.12.1
filelock==3.15.1
Flask==3.0.3
fonttools==4.53.0
fsspec==2024.6.0
importlib_metadata==7.1.0
importlib_resources==6.4.0
itsdangerous==2.2.0
Jinja2==3.1.4
kiwisolver==1.4.5
MarkupSafe==2.1.5
matplotlib==3.9.0
mpld3==0.5.10
mpmath==1.3.0
networkx==3.2.1
numpy==2.0.0
packaging==24.1
pillow==10.3.0
pygame==2.5.2
pyparsing==3.1.2
python-dateutil==2.9.0.post0
six==1.16.0
sympy==1.12.1
torch==2.3.1
torchaudio==2.3.1
torchvision==0.18.1
typing_extensions==4.12.2
Werkzeug==3.0.3
zipp==3.19.2
